+++ interview/forms.py
@@ -12,6 +12,7 @@ from crispy_forms.layout import Layout, Div, Submit, Column, Field
 from django_select2.forms import ModelSelect2MultipleWidget, ModelSelect2Widget
 
 from interview.models import Consultant, Interview, Candidate, Process, Sources, Offer
+from interview.widgets import UploadFilesWidget
 
 
 class MultipleConsultantWidget(ModelSelect2MultipleWidget):
@@ -38,7 +39,7 @@ class ProcessCandidateForm(forms.ModelForm):
         helper = FormHelper()
         exclude = ("anonymized", "anonymized_hashed_name", "anonymized_hashed_email")
 
-    cv = forms.FileField(label="CV (pour une candidature)", required=False)
+    cv = forms.FileField(label="CV (pour une candidature)", required=False, widget=UploadFilesWidget())
 
     helper = FormHelper()
     helper.form_tag = False
@@ -152,11 +153,7 @@ class InterviewMinuteForm(forms.ModelForm):
         model = Interview
         fields = ["minute", "next_interview_goal", "kind_of_interview"]
 
-    # https://docs.djangoproject.com/en/1.11/topics/http/file-uploads/#uploading-multiple-files
-    # Only way to select multiple files is when selecting them at the same time (e.g: using SHIFT + click)
-    document = forms.FileField(
-        label="Document", required=False, widget=forms.ClearableFileInput(attrs={"multiple": True})
-    )
+    document = forms.FileField(label="Document", required=False, widget=UploadFilesWidget())
     helper = FormHelper()
     helper.form_tag = False